refactor: TossPaymentClient 를 컴포넌트로 등록하고 설정 정보 개선을 통해 관심사 분리

+++ src/main/java/roomescape/application/dto/request/payment/PaymentRequest.java
@@ -1,7 +1,7 @@
 package roomescape.application.dto.request.payment;
 
 public record PaymentRequest(
-        Long amount,
+        int amount,
         String orderId,
         String paymentKey
 ) {
